Rename function to generate useRediret to have ReactRouterV5 to distinguish
